backend/src/config.py
import os
from dotenv import load_dotenv

load_dotenv()

SETTINGS = {
    "DATABASE_HOSTNAME": os.getenv("DATABASE_HOSTNAME"),
    "DATABASE_PORT": os.getenv("DATABASE_PORT"),
    "DATABASE_PASSWORD": os.getenv("DATABASE_PASSWORD"),
    "DATABASE_NAME": os.getenv("DATABASE_NAME"),
    "DATABASE_USERNAME": os.getenv("DATABASE_USERNAME")
}