backend/requirements.txt
aiomysql==0.1.1
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
fastapi==0.88.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
mysql-connector-python==8.0.32
packaging==23.0
pluggy==1.0.0
protobuf==3.20.3
pydantic==1.10.2
PyMySQL==1.0.2
pytest==7.2.1
python-dotenv==0.21.1
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.22.0
termcolor==2.2.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
